% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cassavaCropManager.R
\name{datesIn365}
\alias{datesIn365}
\title{function to convert date 1:30/31 to 1:365/366}
\usage{
datesIn365(ddf, hv = FALSE, pl = FALSE)
}
\arguments{
\item{ddf}{is date, in format "yyyy-mm-dd"}
}
\description{
function to convert date 1:30/31 to 1:365/366
}
