% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cassavaCropManager.R
\name{fertilizerAdviseTable}
\alias{fertilizerAdviseTable}
\title{function to put data used in the markdown .rmd}
\usage{
fertilizerAdviseTable()
}
\description{
function to put data used in the markdown .rmd
}
