% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cassavaCropManager.R
\name{fertilizerFunc}
\alias{fertilizerFunc}
\title{function to creat a data frae with fertilizers}
\usage{
fertilizerFunc(urea = TRUE, ureaBagWt = NULL, ureaCostperBag = NULL,
  MOP = TRUE, MOPBagWt = NULL, MOPCostperBag = NULL,
  NPK201010 = TRUE, NPK201010BagWt = NULL,
  NPK201010CostperBag = NULL, NPK151515 = TRUE,
  NPK151515BagWt = NULL, NPK151515CostperBag = NULL, country)
}
\arguments{
\item{urea}{logical tru false if it is avaiable for farmers}

\item{ureaBagWt}{if user define weight of urea in a bag}

\item{ureaCostperBag}{the cost of urea in defined bag
like wise}
}
\description{
function to creat a data frae with fertilizers
}
