% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cassavaCropManager.R
\name{fertilizerFunc2}
\alias{fertilizerFunc2}
\title{function to creat a data frae with fertilizers}
\usage{
fertilizerFunc2(fertData, country)
}
\arguments{
\item{urea}{logical tru false if it is avaiable for farmers}

\item{ureaBagWt}{if user define weight of urea in a bag}

\item{ureaCostperBag}{the cost of urea in defined bag
like wise}
}
\description{
function to creat a data frae with fertilizers
}
