% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cassavaCropManager.R
\name{getFRrecommendations}
\alias{getFRrecommendations}
\title{@param PD: controls not to ask recommendation for dates that are not considerd as planing dates in the zone
 @param rootUP: root rpice for fresh wt. ton/ha it is 270 in TZ and  141 for NG#'
 @param cassUW: Unit weight at which cassava produce is sold, in kg; common measures are 1 (per kg), 50 (per 50kg bag), 100 (per 100kg bag) and 1000 (per tonne); can be NA if user does not know.#'
 @param cassUP: Price of 1 cassava produce unit in local currency; can be NA if user does not know.
 @param maxInv: : Maximal investment in fertilizer, for the area of the field in local currency, NA if FR != TRUE, default = NA (if user does not wish to set an investment ceiling)
 @param rootUP: root price for fresh wt. in ton. It is 270$ in TZ and 141$ for NG
 @param country: "NG" or "TZ"
 @param lat
 @param lon
 @param areaUnits  : Units for area (acre, ha or m2)
 @param area       : Numerical value for area of the field
 @export}
\usage{
getFRrecommendations(lat, lon, PD, HD, country, cassUW, cassUP, maxInv,
  fertilizers, NG_CY_Fertdata, SoilData, area, areaUnits, userName,
  userPhoneCC = NA, userPhoneNr = NA, cassPD, fieldDescription,
  userEmail)
}
\description{
@param PD: controls not to ask recommendation for dates that are not considerd as planing dates in the zone
 @param rootUP: root rpice for fresh wt. ton/ha it is 270 in TZ and  141 for NG#'
 @param cassUW: Unit weight at which cassava produce is sold, in kg; common measures are 1 (per kg), 50 (per 50kg bag), 100 (per 100kg bag) and 1000 (per tonne); can be NA if user does not know.#'
 @param cassUP: Price of 1 cassava produce unit in local currency; can be NA if user does not know.
 @param maxInv: : Maximal investment in fertilizer, for the area of the field in local currency, NA if FR != TRUE, default = NA (if user does not wish to set an investment ceiling)
 @param rootUP: root price for fresh wt. in ton. It is 270$ in TZ and 141$ for NG
 @param country: "NG" or "TZ"
 @param lat
 @param lon
 @param areaUnits  : Units for area (acre, ha or m2)
 @param area       : Numerical value for area of the field
 @export
}
