% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cassavaCropManager.R
\name{sendEmailReport}
\alias{sendEmailReport}
\title{function to send mail from acai.akilimo@gmail.com}
\usage{
sendEmailReport(res = res, userName, userEmail, userField)
}
\description{
function to send mail from acai.akilimo@gmail.com
}
